backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]>=0.24.0
python-multipart
PyMuPDF
Pillow
sentence-transformers==2.7.0
requests
python-dotenv
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
celery==5.4.0
redis==5.0.0
faiss-cpu==1.8.0
paddleocr==2.7.3
paddlepaddle==2.6.0
tinydb==4.8.0

# 테스트용 라이브러리 (기존 파일에서 가져옴)
pytest
pytest-asyncio
pytest-cov
